+++ src/main/java/org/opensearch/knn/index/util/Nmslib.java
@@ -68,4 +68,9 @@ class Nmslib extends NativeLibrary {
     ) {
         super(methods, scoreTranslation, currentVersion, extension);
     }
+
+    @Override
+    public float distanceTransform(float distance, SpaceType spaceType) {
+        return distance;
+    }
 }